Increase the number of rounds if it's a typo
The game has 5 rounds. And a round will end even if the player enters
a typo.

I incremented the value of rounds if it was a typo. This way a round
will only end if the player enters his choice correctly.

A player can enter typos 100 times but the game will not end until
the player enters correct words 5 times.